Moved `DbMock` into `utils` folder for next utilization
