+++ rpgJavaInterpreter-core/src/test/kotlin/com/smeup/rpgparser/smeup/dbmock/C5ADFF9LDbMock.kt
@@ -17,6 +17,7 @@
 package com.smeup.rpgparser.smeup.dbmock
 
 import com.smeup.rpgparser.interpreter.FileMetadata
+import com.smeup.rpgparser.utils.DbMock
 import java.io.File
 
 class C5ADFF9LDbMock : DbMock {
