+++ rpgJavaInterpreter-core/src/test/kotlin/com/smeup/rpgparser/smeup/dbmock/MULANGTLDbMock.kt
@@ -1,6 +1,7 @@
 package com.smeup.rpgparser.smeup.dbmock
 
 import com.smeup.rpgparser.interpreter.FileMetadata
+import com.smeup.rpgparser.utils.DbMock
 import java.io.File
 
 class MULANGTLDbMock : DbMock {
