+++ rpgJavaInterpreter-core/src/test/kotlin/com/smeup/rpgparser/smeup/dbmock/TABDS01LDbMock.kt
@@ -1,6 +1,7 @@
 package com.smeup.rpgparser.smeup.dbmock
 
 import com.smeup.rpgparser.interpreter.FileMetadata
+import com.smeup.rpgparser.utils.DbMock
 import java.io.File
 
 class TABDS01LDbMock : DbMock {
